1.15

+++ src/main/java/net/querz/mcaselector/version/ChunkRenderer.java
@@ -4,5 +4,5 @@ import net.querz.nbt.tag.CompoundTag;
 
 public interface ChunkRenderer {
 
-	void drawChunk(CompoundTag root, ColorMapping colorMapping, int x, int z, int[] pixelBuffer, int[] waterPixels, byte[] terrainHeights, byte[] waterHeights, boolean water);
+	void drawChunk(CompoundTag root, ColorMapping colorMapping, int x, int z, int[] pixelBuffer, int[] waterPixels, short[] terrainHeights, short[] waterHeights, boolean water);
 }
